requirements.txt
fastapi
sqlalchemy
asyncpg
databases
alembic
psycopg2-binary
pydantic
uvicorn
python-dotenv